+++ test/fixtures/templater/jinja_a/jinja.sql
@@ -1,0 +1,1 @@
+SELECT 56 FROM {{ testing_schema }}.{{ testing_table }}